Dostosowanie scenariuszy PyGame do Pythona 3.

+++ docs/pygame/life_str/life_str1.py
@@ -1,8 +1,10 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import pygame, sys, random
-from pygame.locals import * #udostępnienie nazw metod z locals
+import pygame
+import sys
+import random
+from pygame.locals import *  # udostępnienie nazw metod z locals
 
 # inicjacja modułu pygame
 pygame.init()
@@ -19,8 +21,8 @@ pygame.display.set_caption('Gra o życie')
 # rozmiar komórki
 ROZ_KOM = 10
 # ilość komórek w poziomie i pionie
-KOM_POZIOM = OKNOGRY_SZER/ROZ_KOM
-KOM_PION = OKNOGRY_WYS/ROZ_KOM
+KOM_POZIOM = int(OKNOGRY_SZER / ROZ_KOM)
+KOM_PION = int(OKNOGRY_WYS / ROZ_KOM)
 
 # wartości oznaczające komórki "martwe" i "żywe"
 KOM_MARTWA = 0
